MKAudio: restart on device add/remove.
This change makes sure we catch new devices being added and removed,
in preparation of Bluetooth support.

This change can cause an excess of restarts on iPod Touches without
a built-in microphone. (This is the case on any devices where we can
transition from an audio route with input available to one without,
and vice versa.)