MKAudio: handle RouteConfigurationChange in AudioSession route change callback.
This fixes the erroneous infinite restarting bug, which is seen
when running 'Mumble for iOS' 1.2.3 on iOS 7.1.

Something with the VPIO AudioUnit is causing the route change reason
kAudioSessionRouteChangeReason_RouteConfigurationChange to be emitted.
This seemingly happens when the VPIO AudioUnit is initialized, or
somewhere around it.

The MKAudio code didn't handle this change reason (it was introduced
in iOS 7.0). Because MKAudio didn't handle it, it was treated as a
'restart' reason, and thus MKAudio would restart both MKAudioInput
and MKAudioOutput when it received the RouteConfigurationChange reason.
This restart triggered the same change reason to be emitted, leaving
MumbleKit's audio subsystem in an infinite restart loop.