Fix a bug with buffer size calculation in the AudioUnit inputCallback
This commit fix an error in the formula used to check if the buffer is too small.

Moreover the AudioUnitRender function modifies the mDataByteSize members with the actual read bytes count. This makes the buffer continuosly reallocated even if not necessary. This commit reset the mDataByteSize member to the actual buffer size.