MKCryptState: fix initialization of CryptState class.
Previously, the constructor of CryptState was not called
because CryptState was a member of MKCryptStatePriv, which
was allocated via malloc.

This gets rid of MKCryptStatePriv, and instead just keeps a
pointer to the private CryptState implementation directly in
MKCryptState. As a result, we now allocate a new CryptState via
'new' in init, and clean it up in dealloc via 'delete'.

Fixes mumble-voip/mumblekit#31